Dish Today is an app that is built by mainly considering the  household women in mind. The main aim of this app to generate the combination of dishes as per the region selected by the user and user have to only select are deselected the suggestions shown by the app. Not only this the app has 3 more feature into it which are discussed in the objective of this synopsis.
 Technology used in this app are Android studio 4.1, Maven Repository, Android
Profiler (Memory usage, CPU usage, Network Energy), SDK and ADB, IntelliJ,
Firebase Debugger feature in Mobile.

The app also has Blogger Section in which the user can put the snap
and the fellow user can like it and comment on it, user can ask for recipe through
Direct message. The user can also search for new recipe through our search bar
which will redirect the user to YouTube with the same input. Our Homemakers
are so good in cooking food but don’t get the platform to showcase their talent so
we are thinking to monetize our blogger section in Coming future one the app
will gain the popularity.